requirements.txt
streamlit>=1.28.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
